Protocol change: 16 bit annotation and layer IDs
The first 8 bits of the ID is the user's context ID. This allows each user
to make up new clash-free IDs without the server's guidance.
Predictable IDs are required for the local fork to work correctly.
Besides simplyfing the server a little, this also makes it possible to
implement features such as "paste to a new layer".