Fix retconning of indirect paint mode
In indirect mode, all pen moves are applied to a private layer first
and composited on pen-up, meaning only the relative order of
the pen-up commands matters. This is the opposite of direct mode,
where pen-up has no effect and the individual moves matter.

TODO (visual glitch): if two indirect strokes are retconned,
the pixmap cache will not be updated. This is a not a critical
problem, since the cache will get refreshed eventually, but it
can still be visually distracting.