Add hybrid tablet event mode
QTabletEvent reports invalid coordinates (constant or random offsets) on
some tablets/drivers. This commit adds a hybrid tablet mode where the
pressure value is read from the tablet event, but the mouse event is
used for actual drawing. This mode is enabled by checking the
"Enable bug workarounds" checkbox in the settings dialog.

Relates to issues #177, #152, #151 and #23