Offer smoothed points as soon as any data exists
The implementation of smoothPoint() already allows for the buffer to be
partially full, so this just changes hasSmoothPoint() to return if
the buffer is non-empty. It is currently never called in such a way
that this should return false, but since smoothPoint()'s return is only
defined if there are point data, it seems sensible to keep the API to
determine this.

Seems to be causing some small 'hooking' artefacts at the start of
strokes occasionally, but I cannot tie down the cause. The actual
smoothed data seems fine. It may be due to the effective ramp-up of the
smoothing level as more data is available to work with.