Prefill the smoothing buffer with the first point
This changes the blending of early points; rather than the first
point's influence falling off as 1/n, it falls off linearly. This is
what Krita does and it seems to help mitigate the angle irregularities
near the start of the stroke.

_count is now a bit meaningless as an integer since the buffer is
either full or empty, but leaving as-is since that may change back.