Add initial Travis-CI configuration
Tells Travis to try to build commits with G++ on Linux.

Because Travis build nodes are Debian-based and this is just getting
dependencies through apt, doesn't configure with all features enabled
and won't work on 2.x series branches (so building master is
disabled, unfortunately). Even then it takes a lot of PPAs to get it
up to date, and then a little hackery, but at least all that is
isolated only to Travis builds.

Clang building is disabled due to it failing for reasons I haven't
investigated yet (can't find system 'algorithm' header).