Remove kicked users from the session immediately
Even though the client won't be actually disconnected until the send
buffer is cleared, it can still be removed from the session.
This prevents multiple kick commands being logged for a single logical
kick event, when the user is not immediately removed from the list
and the operator clicks on the kick button several times.