Set context IDs on messages to be sent
Now that context IDs are not forced in client::sendMessage,
we must make sure to set the correctly when constructing the messages
in the first place.

This doesn't matter when using a real server, since the server enforces
context IDs, but loopback mode will not work correctly if the IDs
are not right.