Remove "persistent session" box from host dialog
Whether persistent sessions are available is not known until connecting
to the server, therefore checking this box often has no effect, leading
to confusion. After logging in, the persistence checkbox in the session
settings dialog is either enabled or disabled, depending on whether
this feature is enabled on the server.