Emit progress at the end of saveNextFrame
The slot the progress signal is connected to may spin the event loop
(which, for example, progress dialog's setValue does,) in which case
saveNextFrame can end up being called recursively before the
frame counter is incremented. We can avoid this by emitting progress
at the very end of the function.

Issue #563