Don't check if op is fully caught up
This check's purpose was to prevent operators who have just
joined the session and haven't fully caught up yet from autoresetting
the session. However, the logic doesn't seem to work right, since in normal
operation it's possible the user is just a message or two behind and the
request is therefore never sent.