Automatically check for new versions
Fetch an appdata file from drawpile.net and check if there are
any newer versions listed. Unless disabled, this is done automatically
on program startup and can be triggered manually via the help menu.

If a new incompatible version has been released, a notification
is shown in the join dialog.