Add colorwheel shape, angle and space options
A new tab in preferences allows you to set these and wires them up to
the matching options in the ColorWheel class. If no settings exist yet,
it'll default to a rectangular, fixed colorwheel with a square shape, as
before.

From my testing, the Lch color space acts in a really confusing manner.
I guess this would make sense to people who know how to work it, but I
dunno if it's actually a use case for Drawpile, so maybe it could be
removed to avoid bug reports by users that are as perplexed by the
behavior as I am.