Make input configuration dialog look more proper
Adds a few things to make it look more reasonable.

Sensible padding around the edges and between its elements, in line with
other dialogs.

A minimum size for the bar on the left so the handle doesn't get
clipped on the left.

Expanding vertical layout policy for the curve. Otherwise hiding the bar
on the left will cause it to collapse.

A close button at the bottom. It would be more in line with other
preferences to have this be a set of OK and Cancel buttons instead, but
since you probably want to be trying out your input settings while
you're adjusting them, this is probably fine.