Implement canvas viewport entry hack for Linux
On my Linux system, when entering the canvas viewport via tablet, the
cursor doesn't update properly. This implements a workaround that allows
Qt to generate synthetic mouse events for tablet move events, which
function properly. The canvas view filters out these synthetic events
when the hack is enabled, which means that there should be no change in
behavior.

For reference: this is Arch Linux with whatever current Qt5 it comes
with, the issue has been there since forever I think. This might be
worth raising as an issue with Qt itself, but that'd need a more minimal
example to show off the issue in isolation.