Implement a layer properties dialog
Krita, GIMP et cetera also have a dialog like this, to have all the
information there to edit at once.

The opacity slider also has a spinner in this view, to allow setting a
precise value and not having to twiddle slider positions until they
match up between different layers.

This replaces the in-line layer title editing. Together with #971, this
finally fixes #593, since it's the in-line edit control that's causing
the problems. Moving it to a separate dialog removes the issue
altogether. The dialog is set up so that the amount of user input for
renaming a layer remains the same: double-click, type a new name and hit
return.