Port document loading to Rust
Remove the CanvasLoader class. This was a vestige from earlier when
local canvas history was used to initialize session history. (Now, a new
snapshot is always generated when needed.)
Therefore, there is no need to go through the intermediate step of
converting an image to Drawpile commands, then executing those commands.
Instead, we can prepare a LayerStack and just replace the existing
canvas with the new content.