Make classic brush lookup tables consistent
There's 256 different hardness values, but only 101 lookup tables for
them. Before this commit, the lookup table was calculated using the
hardness value gotten from calculating `hardness / 255.0`, which was
then assigned to the lookup table at index `int(hardness / 255.0 *
100.0)`.

That's inconsistent however. For example, hardnesses 227, 228 and 229
all get assigned to index 89, but calculate slightly different values
for the lookup table. That means, depending on which order you draw dabs
and how the cache behaves, you get different results.

This commit now always calculates the hardness according to the index of
the lookup table, so index 89 will always contain the value for hardness
0.89 without any jitter.