Make layer resize color sampling consistent
When resizing a layer, Drawpile tries to guess a color to fill the
background with. For that, it uses a hash table to keep track of the
frequency of color samples from the edges of the canvas and grabs the
color that has the highest frequency at the end.

However, in the (exceedingly theoretical) case where two colors have the
exact same frequency, Drawpile will effectively pick whichever color is
the first one it finds in the hash table. QHash has no defined order, so
it picks at random. This could lead to inconsistent colors being chosen
for different users or when playing back a recording if the hash table
ends up in a different order.

This commit adds a tie-breaker for those cases: it uses the color with
the smaller numerical value. That way, the resulting color will always
be consistent.