Add communication between Lua and the browser
From Lua to the browser: CustomEvents with type "dpevents" are sent to
the window object. They contain a JSON string which decodes into a list
of [type, payload] pairs, as well as a pointer (that is, a number) to
the Lua state that is to be passed back to the application.

From the browser to Lua: using the function _DP_send_from_browser,
events are stashed into the Lua state's client event queue. From there,
they're dequeued to Lua as normal.

The application sends a DP_APP_INIT event when it starts up and a
DP_APP_DISPOSE event when it exits (which really shouldn't happen in the
browser, you'd just close the page instead.)