Implement retrieving change bounds
At least on a per-tile basis. This is different from how Drawpile does
it, which keeps track of the actual change bounds using separate calls
to update them. Kinda complicated to get that working in a threaded
fashion though, while not gaining much from it, since multiple people
drawing on the same tiles but not quite the same pixels is a
pathological case to begin with, so I chose not to.

Used by Drawpile's conflict resolution.