Implement annotation list model diffing
To figure out what changed between two annotation list states. Checks
what changed between them and peforms callbacks to remove, add, move and
modify elements to turn the old state into the new one. The intent is
for this to be used with GUI models like Qt has, which prefer this kind
of explicit change rather than resetting the whole list.

The implementation is reasonably generic and should be usable for layer
property lists as well.