Don't composite onto document background
The document background shouldn't be treated as if it actually existed,
since that displays results very much unlike they would show up when
exported as a flat image without the background. This can be seen very
obviously with the Erase blend mode: I expect to see the checkerboard
pattern behind the erased area, I don't want to erase the checkerboard
itself. The same also applies to blend modes like Multiply: they
shouldn't multiply the checkerboard, because when I export the file,
there's nothing there to multiply either.

This also gets rid of the background tile cache, since that becomes
superfluous due to this change.