Smooth movements of usermarkers by default
MyPaint brushes can cause dabs to appear in wildly different places,
causing the user marker to jerk around like crazy. This commit makes the
markers interpolate their position instead, meaning that they're less
accurate, but also way less annoying.

Laser pointer markers still work as before, since their accuracy is more
important and they don't have the crazy dabs issue in the first place.

This speeds up the animation timer tenfold, since otherwise the movement
is too inaccurate. It's also definitely a hack, but it seems okay to me,
the position of the user markers is still close enough. Maybe it could
be restricted to only to the interpolation when a MyPaint brush is used,
but that might be too much effort for something that's barely an issue.

What I definitely would *not* do is introduce new messages to report the
cursor position, the markers aren't worth that.