Support colorize and posterize for MyPaint brushes
But no spectral painting support, because that's a pretty weird feature
with performance implications. Don't wanna do a lot of matrix math with
every brush dab.

This breaks protocol with the previous beta release, since the MyPaint
brush dabs now include colorize and posterize information. Maybe we
should increment the protocol version accordingly, even if it's just a
beta.