Update canvas visuals after layer merge properly
The code assumes that merging a layer doesn't cause a visual change, but
that's not true for almost all blend modes. For example, when I merge a
Multiply layer, all pixels that multiply with transparency will
disappear.

This commit fixes that: when the blend mode of the source layer is not
Normal, we use the area of effect of the source layer to update the
canvas properly.