Add protocol generator from Drawpile 2.2
Drawpile 2.2 generates its protocol message implementation using a
definition file and some Python scripting. We do the same now.

This commit copies protogen.py and protocol.yaml from Drawpile's
sources. It amends the former by a copyright header. In the latter, the
ServerCommand's msg field type is changed to a string because it's
really a JSON string and the MoveRegion message is reintroduced, since
Drawdance already has an implementation for it, so we might as well keep
it wired up.

Protocol generation can be done using the one-stop script
scripts/protogen. It will automatically set up a Python virtual
environment, install the necessary dependencies, run the generator and
format the generated code.