Implement document metadata
Storing it in the canvas state and handling the metadata messages
accordingly. String metadata doesn't exist in Drawpile, so the message
to set a string metadata field just bails out for now.

Also adds a test for handling these kinds of messages, consolidating the
common code with the annotation handling test.