Move view mode logic into its own file
So it can be re-used for rendering as well as for the local paint engine
view in the form of a view mode filter. I think the logic used now is
also more correct when dealing with manual frame modes, since it ignores
all invalid layer ids in the frame.