Allow introducing artificial lag into the client
To help reproduce issues that happen on bad connections. Works better
than the randomlag parameter on the server, since that doesn't really
simulate client lag well and can be toggled at runtime. This also
introduces a new Developer Tools submenu in the Tools menu and puts the
Profile option in there. Fixes the issue with the Profile action not
updating properly too, since now it's updated when the menu is opened,
so it'll always be up to date.