Rework brushdock UI so it can be made thinner
Making the blend mode box the separator and giving it a small minimum
size so that it doesn't force itself to fit every blend mode name. This
also fixes the UI jerking vertically when switching between classic and
MyPaint brushes, since we now retain the blend mode box's vertical
space.