Don't resize the window when hiding docks
Turns out my tiling window manager was hiding that fact from me. Now we
save and immediately restore the main window state to make it update
immediately, that makes it work properly without any flickering or
resizing. It's a hack to do it like this, but it works, unlike
everything else I tried.