Rework flood fill settings user interface
Changing the checkboxes into combo boxes with less confusing and
actually translatable terminology. Sample Merged is now a selection
between Current Layer and Merged Image. The Fill Under and Erase boxes
are now a selection between Normal, Behind and Erase, which lines up
with how they actually work and is consistent with layers and brushes.
This also removes the rather odd restriction that you can only erase if
the sample source is the current layer.