Don't render unnecessary checkerboard background
It's very rare in practice for the checkers to actually be visible. Only
if the canvas background isn't opaque or there's an Erase layer at
top-level (or inside a pass-through group) is this even possible. We now
track this possibility and only bother to render the checkers behind the
background if they could hypothetically be visible. Since this affects
every single visible tile, it improves performance by a good bit.