Make dpconv functional
It is now able to take dprec and dptxt recordings and turn them into
ORA, PNG or JPG images, as well as converting them to dprec and ora
recordings. This commit also consistenizes the arguments given to
DP_save and DP_player_new so that they don't rely exclusively on
guessing the type by file extension or magic.