Add a FILE_IMPL build option for Qt files
Changing it from STDIO to QT will go through Qt's QFile and QSaveFile
instead of using stdio. This is necessary on Android because it uses
weird URIs instead of actual paths, which don't work properly for fopen
and such. Also has the advantage of QSaveFile supporting writing to a
temporary file and then renaming it to the target file on success, which
is something that was previously manually implemented in Drawpile.

Also changes some build options from taking QT6 as an implementation
option and instead adds a separate QT_VERSION option that can be set to
5 (default) or 6, since it really doesn't make sense to use multiple Qt
versions in the same program.