Don't create an annotation when switching tools
The annotation tool didn't reset its currently selected id, so switching
to a different tool made it think that the user just let go of a held
left-click and would create a new annotation. This is now fixed, it
properly resets the selected annotation id to 0 after letting go of a
button.