Empty out master branch, add redirection notices
Since the branch has been in an unchanged state for really long and
scripts are likely to have grown dependent on it, I'd rather not break
them in a weird way by replacing this with the 2.2 dev branch. So now
this here is an empty branch that just contains notice files to go look
in the correct branch.