Implement a Krita-esque stabilizer
Instead of the MyPaint-like one, because it's just not that good. This
one is time-based, so it doesn't stop dead in its tracks when the curser
stops moving, and it can manage to finish up strokes. This gives every
canvas point timing information, which is why it touches so much. Also
passes the view zoom into the brush engine now, since that's something
MyPaint brushes actually use as an input.

Also now passes the view zoom to MyPaint brushes, since the signatures
of the functions changed anyway.