Make KisTablet work without patching Qt
Because KisTablet's events are non-spontaneous, while Qt's native events
are, so I think we're able to just filter them according to what kind of
input mode we have enabled. Also patches around on KisTablet a bit to
allow switching the input mode without requiring a restart of the entire
application and adds the HAVE_KIS_TABLET define again, it went missing
in the mergepocalypse.