Always reset Qt6 to Windows Ink mode
When switching to another input mode, rather than just leaving it at the
mode it was at. Leaving it set to Wintab mode causes menus to not react
to tablet input anymore, but resetting it to Windows Ink makes them work
fine, even in conjunction with KisTablet's Wintab input mode.