Get rid of the automatic timeline
Because we don't need two animation systems fighting with each other,
one is plenty. The automatic timeline was only partially implemented
since Rustpile anyway, with no support for fixed layers. If there's
really a clamoring to bring it back, it can be in the form of a button
that auto-populates the timeline or something. We still support fixed
layers on the protocol level, so it's just a client-side matter of
interpreting it.