Block frontend updates during a canvas reset
For one, we now block updates from the issuing of a reset to the first
canvas resize message, since during that period we just have a null
canvas and it's not useful to show that. More significantly, locking
user 0, which before was a nonsensical operation, is now given meaning
to introduce a reset lock that stays active until the next UserAcl
message clears it again. While this lock is active, all updates from the
paint engine to the frontend are suspended, meaning there won't be any
flickering and you won't get thrown around to different layers as the
list repopulates.