Fix Windows build time asset management and unity build options
Copying the entire assets directory does not include i18n bundles
so is not enough to solve missing files for Windows developers.
It is also slow, since it copies hundreds of theme files. Finally,
it *must not* be done on every platform; on at least macOS, this
change caused assets to be dumped into the wrong directory and
created a screwed up application bundle.

Android already required this kind of asset management at build
time so this patch just reuses that existing functionality.

CMake’s `UNITY_BUILD` is a global property, not a variable, so
there is no conflict to using it as a variable, but actually
the idiomatic way to enable unity builds is by setting the
`CMAKE_UNITY_BUILD` variable. Since ‘global’ variables are actually
per-project, disabling it for incompatible sub-projects is simply
a matter of forcing it to `off` in the subproject CMakeLists.txt.