Make system palette and style work properly
By remembering which was set on application startup. Linux doesn't have
a "" theme. Windows' default palette isn't the empty palette. This takes
care of both of those.

Removes the special handling of the system style on Windows, since we
now offer the user a choice between classic Windows and Vista styles.