Clean up the state of translations
Brings Spanish translation back from 2.1, not sure where that was gone.
Sticks the German translations I did back into the current state.
Deletes the unused drawpile_*.ts files in the libclient directory.
Renames the translation files to contain both the language and country,
not just the language. Adds Spanish and Norwegian to the supported
languages and excludes the latter from warnings about missing Qt
translations since it's the same situation as with Vietnamese.