Clean up and add NSFM option to host dialog
Instead of always guessing the NSFM mode from the title, it now provides
a checkbox for it. Typing a title that looks to include an NSFM keyword,
the box will be ticked automatically. If parental controls don't allow
joining NSFM sessions, the NSFM option will not be visible and the OK
button will be disabled.

Also removes the private listing option, since it's not useful anymore
and also pretty confusing as to what it does. And replaces the group
widgets in the UI with a few separators.