Allow turning off automatic checking for updates
With a new checkbox in the network preferences. On first run, we also
don't immediately check for updates now, instead presenting the user
with a greeting message. If they want to turn off automatic update
checking, they can do so in the preferences. Not having automatic
updates enabled will cause existing news to go stale after three days,
prompting them to either turn on automatic checking or to check for
updates manually.