Make the color picker use the old palette widget
Because it just works better for the purpose than QtColorWidgets'
swatch. The widget has been adjusted to use their palette though, as
well as making the painting more consistent with the chosen style.

Also adds a "pick from screen" button that's really convenient for
picking colors from reference images, so it's nice to have direct access
to it instead of going through the color dialog. This also moves the
buttons and size slider to the header widget, since we got room there
anyway.