Make recent hosts use the same mechanism as files
Writing into the SQLite database instead of using the config file for
its purposes. The RecentFiles class has been renamed to Recents to
reflect this change. Join and host hosts have been unified into a single
table, the entries in which remember if they've been used for hosting
and/or joining so that a sensible default can be chosen for the host
dropdown.