Avoid colors tending toward black when smudging
By adding a fudge factor. Not too sure why it happens in the first
place, since we round things properly, but this seems to be a fix for
anything going wrong in that regard. Maybe just an artifact of the
algorithm or caused by the inaccuracy of the 8 bit protocol that we have
to compensate for.