Render canvas asynchronously
Instead of hanging the main thread for it. This should prevent the UI
from chugging on a busy canvas. It especially improves the situation
when the navigator is open, since it no longer causes huge lag spikes
from synchronously rendering the entire canvas every few seconds. But
also other large-scale operations like filling an area or changing layer
visibility now don't lock up the whole application, instead they render
in parallel.

Tiles within the current view are prioritized and rendered first.
Anything outside the immediate view is either rendered afterwards or not
at all, depending on if the navigator is open or not. To avoid tiles
appearing to "glitch" at the edges of the screen, the rendering is still
synchronous when the view is manipulated. That means the application
still chugs when doing so, but that feels better than a bunch of stuff
hastily flickering into being.