Handle canvas resizes while transforming
In particular, handling offsets properly when the canvas is extended
left- or upwards. It now properly offsets the source region accordingly.
Previously, it would keep it in place, causing the result to appear in
the right place, but the source wasn't from where the user had actually
put their selection.