Fix wheel going gray when changing color space
This was caused by an attempt to convert the hue to the current color
space, which may be an indeterminate value if the current color is fully
desaturated, causing Qt to return -1. Hue is the same value in all three
color spaces though, hsvHueF just calls hueF under the hood, so this
commit just removes the hue modification when changing the color space
altogether.