Allow mirroring the inner color wheel selector
Most color selectors in other software, such as Krita, GIMP, MyPaint and
many others, go from least saturated to most saturated. QtColorWidget's
selector goes the other way round, which is valid, but unusual. This
adds the ability to mirror the inner selector, giving the option to put
it in line with other software.