Parse emoji manually instead
Because PCRE doesn't want to play nice on Windows and the regex has some
annoying false positives. This new way is basically a manual regex,
generated from the Unicode definitions of all possible emoji. It doesn't
use any Unicode properties or anything, instead it just walks UTF-16
characters. It doesn't work on big endian, but we don't build for any
platform of that sort anyway.