Barely implement a "mobile phone" UI mode
Makes the application usable on a phone without having the text be
ginormous, squashing everything into garbage.

Adds on-screen HUD controls to show docks, chat and toolbars. There's
still some wonk here, but it's good enough for now.

On Android, the initial UI state is guessed based on the screen
resolution in pixels and millimeters. This detection may not be too
terribly accurate, but the preferences have a toggle.

Font size is now actually set to a valid point size on Android, rather
than leaving it at the default that causes some other issues because
it's only a pixel size. This can also be adjusted in the settings in
case it's actually *too* small or something. May also be helpful on
desktop devices with a really low resolution.