Exclude builtin server along with server component
When not building the server (-DSERVER=OFF), the builtin server now
isn't included either, meaning hosting on this computer won't be
available. That means on Android, the builtin server is never available.
Which makes sense, since Android likes to suspend applications, so it's
not particularly suited for running a server on it.