Show busy cursor when waiting for reset dialog
Building the reset image takes some time, but there was no indication
that anything was actually happening. Now we show a busy cursor so that
it's clear that Drawpile is doing something.

Also introduces a convenience class to do RAII-based cursor overriding,
which is now used in all places where it makes sense, instead of doing
manual setting and restoring of the cursor.