Use normal Qt layouts for preferences dialog
Instead of the custom SanerFormLayout, because it just doesn't work
properly. On Windows, the dialog gets stuck at a minuscule size if you
have multiple monitors. On right-to-left languages, everything gets
mangled. On Android, the screen may not fit the dialog vertically. Those
issues make the dialog effectively unusable, so the minor advantage of
some slightly nicer layouting and looking slightly more native on macOS
isn't worth it.