Allow indirect mode without opacity pressure
In Drawpile 2.1 sessions, because that particular case actually works
fine and is the same when using Normal and Alpha Darken blend modes, so
there's no desync or anything. It's not what indirect mode is supposed
to be for, but some people use it to that effect, so we should continue
to let them do so.