Implement Photoshop Document (PSD) saving
Because apparently plenty of closed-source programs don't support ORA
files, despite them being so trivial. Considering that the PSD "spec"
(if you can call it that) is available publicly since a few years ago,
there's no licensing issues anymore. And it's also a good way to
implement a larger feature in Rust, and file mangling is something it's
particularly well-suited for. The format is still a terrible mess, but
the implementation is pretty okay this way.