Implement loading of PSD files as well
Using psd_sdk, which is a C++ library that seems to parse things pretty
well, unlike the Rust psd crate. It also had a bunch of writing code and
such, but I threw out everything that wasn't needed. There's also some
patchery to actually make this compile cleanly and to not use wchar_t
for file names.