Implement Roles tab in session settings
Shows operators a list of registered users that the session is aware of
and allows setting their operator and trusted status even when they're
not online. The list can be exported and imported again to allow for
restoration of a group of known users without having to re-op and
re-trust them as they join.