Properly initialize color palette internals
Because uninitialized values keep triggering a ubsan warning on every
startup. It now properly initializes `dirty` and `columns`, since the
member functions called in the constructor want to read those values. It
also now always sets `dirty` to `false` at the end of the constructor,
it was doing that for all except one of the constructor variants.