Exclude va_list and FILE from bindgen
Because they cause really funky code to be emitted that is totally not
valid on any other platform than the one that generated the bindings.
Also allows us to get rid of the allowance for non-uppercase globals and
non-snake-case functions, since our C code complies to those. It doesn't
work with non-camel-case types though, due to our DP_ prefix.