Relax transform size check
It was making sure that transforms didn't exceed the canvas size, but
that can totally legitimately happen when just transforming a layer that
covers the canvas. So now it allows up to twice the canvas size in
either direction. Previously, you had to remember to cut and paste your
selection to translate it from a transform to a paste command, which is
a silly workaround.