Fix infinitely recursive signals in color dialog
When using HSL or LCH color models. Changing the saturation when the
color doesn't actually have a saturation would end up setting it
alternately to 0 and the chosen value, bouncing signals between spinners
and sliders back and forth until the stack overflowed. This removes a
duplicate signal from spinners to sliders, since changing the spinner
already sets the slider in setColorInternal, no need to do it again.

This means that trying to change the saturation in HSL and LCH mode when
black is selected will keep the value at zero. That's still not very
nice, but better than bringing the whole application down.