Add notes messages to host dialog
When a title is missing, it shows a message to that effect.

When no password is set, it warns the user that their session will be
public, because some users don't understand that this is what's required
to prevent strangers from joining their session. So they just kick
anyone who joins, causing grief.

When the user chooses to host on their computer, they will get a warning
about it requiring further setup and a link to a help page that tells
them how to deal with port forwarding.