Don't reset annotation format when there's no text
Qt inexplicably obliterates the current text format when the document is
blank, so we re-reset it to fight against it now. This avoids nonsense
like the font becoming some unreadably tiny font if you clear the
annotation text or your font changes getting eaten if you set it before
you start to type.