Don't lose focus when typing into brush sliders
It happened because they got hidden and then re-shown when determining
visibilities, which caused them to lose focus in the process. That was
done so that they don't expand the layout, but now they just disable
updates while manipulating visibility instead, which solves the problem
as well and doesn't cause any focus changes.