Add keep-alive messages from the server
To help avoid the client thinking that it timed out if its upload queue
is too full to get a ping out. During login, the client reports that it
understands these messages, which will cause the server to send a new,
blank control message of type KeepAlive if it didn't send anything else
in the last 30 seconds. The client won't do anything with it, its
purpose is just to reset the idle timer. Since these messages are only
sent conditionally, the protocol doesn't get a bump.