Use a single palette for GIF exports
Instead of a new palette for every frame, because that introduces
inaccuracies and leads to annoying flicker. The palette is now generated
from the merged image, which may not always be ideal, but the user can
influence it by putting some kind of non-animation layer on top of their
frames if necessary. Since a single palette allows it, we also now use
cumulative frames, deduplicating pixels that were identical to the
previous frame. This change also has the nice effect of making GIF
exports way faster and the files to be significantly smaller.