Allow using eraser tip to toggle eraser mode
Lets the user configure if they want their eraser tip to toggle between
the brush and eraser slot or if they want it to toggle the eraser mode
of the current brush (or to ignore the eraser tip entirely.)

Also adds compatibility with Emscripten, where the eraser doesn't report
if it's brought near, instead it draws with button 0x20. This has
slightly different behavior in that it doesn't toggle the erase mode,
since that would flicker back and forth every time the pen is put down
and released, but instead just overrides the blend mode with erase.