Allow setting a server-wide session max user limit
Rather than having it be always at 254. The limit will apply to any
newly hosted session and regular users can't raise it higher.
Moderators, web admin API and session templates aren't bound by this and
can raise it up to 254, like before. After it has been raised above the
limit, users are allowed to reduce the user count to lower values above
the limit, but can't raise it back afterwards.

For example, the session user count limit is 32. A user can't raise it
above 32. A moderator raises it to 64. A user reduces it to 50, which is
allowed, because it's lower than before, even though it's above the
server limit. But they can't raise it back to 64 afterwards without
moderator intervention.