Only ask user to confirm self-signed cert changes
Since that's the only case for which it makes sense. Regular
certificates get renewed regularly and there's no user intervention
necessary, those certificates are already trusted by the root CA. If the
user really wants to pin such a certificate, they can still mark the
host as trusted. We should probably rename this to certificate pinning
to reflect what that actually does.