Use static getSaveFileName function on non-Android
Because Windows spews a bogus warning if a nonexistent file to select is
given for a regular QFileDialog, but this static function handles it
properly. The dialog is still needed on Android because various pieces
about the native file picker dialog behave really strangely, like the
file name being taken from the title.