Load settings defaults after static initialization
Because on Windows in debug mode some enums tend to crash during static
initialization if an attempt is made to load them that way. So now the
default values are behind a function instead.

Of course that causes the next issue of QSettings being unable to load
enum values of types that have never been instantiated. To resolve that,
we now call every default value function once before instantiating the
settings object, which initializes all the metatype stuff.