Add dual color button
In the edit tools, since that seems to be the only sensible place for
it. It works much like it does in Krita and GIMP (although the code is
original), showing the foreground and background color and providing
four spots to click. Foreground and background color square will open a
color dialog to change it, a double arrow at the top-right will swap the
fore- and background color, a reset button at the bottom-left will reset
the foreground to black and background to white. All of these functions
can also be bound to keyboard shortcuts.

This also gets rid of the rather strange secondary recent color swatch.
Swapping colors now swaps foreground and background, rather than
toggling between a set of swatches.

The background color isn't used for any tool currently, but it can be
used to buffer a color there.