Make fill source a two-dimensional setting
Instead of having a drop-down with all the options, it now provides a
set of checkable buttons to pick the kind of source: merged image,
merged without background or layer. The drop-down is only activated when
picking the layer mode and continues to provide the options of current
layer and any other layer in the canvas.

This makes the interface less annoying to use because you don't have to
deal with the unwieldy dropdown at all, you can just right-click on a
layer to set it as the fill source and go to the merged options by
clicking on them directly. It will also now remember which button you
checked.