Allow switching annotation rendering
Between vector (the old way, floating vector text), smooth (anti-aliased
text rendered to a bitmap) or pixel (aliased text rendered to a bitmap.)
The latter two are much more accurate for text intended to be merged,
whereas the first one still makes most sense for rules text or whatever
that's not really part of a drawing.

These properties are transmitted as flag bits in the protocol. Aliased
text is the alias flag, rendering to a bitmap is the rasterize flag.
They are saved to and loaded from ORA files accordingly. Having it split
into two flags gives a fourth option, namely aliased text not rendered
to a bitmap. Aliased vector text is pretty useless and confusing ("it
doesn't alias anymore when I zoom in") though, so this option is not
exposed in the UI.