Allow switching interface mode dynamically
The switch between desktop and small screen mode can now be done at
runtime and no longer requires restarting the application to apply.

There's also now a "dynamic" option, which will automatically change the
interface mode based on window size. This is enabled on Android and in
the web browser. On desktop platforms, it defaults to desktop mode, as
you would expect.

Implements #1228.