WIP: Add pixel rulers
This adds pixel rulers to the top and left sides of the canvas. This is
only implemented for the new renderer, as the previous renderer is going
away eventually.

This handles auto scaling the division spacing and font size, and uses
the current theme's colors for drawing the ruler.

Current issues:
* Not very pretty, would be nice to have someone mock up a better look.
* Does not support canvas rotation. Currently just turns off when the
  canvas is rotated.