Intuit error cause when WebSocket connection fails
Because there's various ways that this can go wrong, with the browser
providing zero clue why the connection failed. We check for inofficial
clients, because many servers have an origin check for web.drawpile.net,
non-wss:// URLs, trying to connect to an IP address instead of a proper
domain name and finally just punt to telling the user that the server
may not support joining via browser. This message is only shown if the
connection fails without receiving a single login command.