Disable Qt6 image reader memory limit
It's set to 256 MiB by default, which is not enough to load large
pictures. Qt5 didn't have this limit at all, so we're just turning it
off altogether. If someone really wants a limit for some reason, they
can still set the QT_IMAGEIO_MAXALLOC environment variable, which will
act as an override.