Unbreak MSI install scope that CMake 3.29 broke
They changed the default scope from an arguably broken halfway
installation between per-user and per-machine to a proper per-machine.
Problem with this being is that it utterly annihilates the ability to
update, because the MSI will see that the existing installation is in a
different scope and then, instead of doing something sensible like
overwrite the per-user with a per-machine install, jams a whole jumbo
box of crayons straight up its nose until its cranium is filled with a
colorful slurry, then proceeds to install the application again but not
overwrite any of the applications files. Which means that the user will
have two desktop shortcuts, two start menu shortcuts and two entries in
the add/remove programs control panel, both of which point at the same
application, which is still at the old version. Running the installer
again and picking "Repair" will then proceed to overwrite the files, but
you'll still be seeing the application twice.

Since the old way the installer works only causes some minor issues for
people who have multiple user accounts on the same machine (I'm sure
there's dozens of them), we revert to the old behavior for now. That's
deprecated, but whatever.

Fixes #1201.