Add alpha darken lerp blend mode, not used yet
This blend mode is closer to what Krita uses for indirect mode. It makes
brushes a lot less soft, but avoids most of the weird halo effects you
get when overlapping them.

This blend mode is not currently used, since doing so would break
compatibility. But it will probably become the default indirect mode in
2.3.0.