Keep selection operations local
Instead of sending them over the network and making all users execute
them. They now instead send LocalMatch commands across the network,
which the client will match with the appropriate commands from the local
fork. Any LocalMatch that reaches the canvas state has no effect.
Selection masks aren't sent along with the command to save space, only
the size of the mask is, which is good enough for matching with the
local fork.

This has various advantages. Other clients don't get bogged down by
having to process huge selection masks from other users, catching up
after a disconnect doesn't process a bunch of pointless masks again, the
session history isn't filled up as quickly and if a selection causes a
crash somehow, e.g. because the outline runs the client out of memory,
it no longer gets restored upon reconnecting just to crash them again.
We also don't need to do as much compatibility hackery around PutImage
anymore, since now the only command that needs that treatment is the
LocalChange one, which allows removing a bunch of compatibility code and
concentrating it on the handling of this one message type.

This doesn't obstruct the plan of adding brush masking with selections.
We can just send the parts of selections that are needed to mask the
brush at the instant they are actually needed. Since we have multiple
selection ids per user to work with, we can do so without causing any
conflict with the currently used "main" selection with id 1. That kind
of system would also be usable for a "stay within the lines" brush mode,
where we'd generate an ad-hoc mask in the spot where the user is drawing
and send that along using other selection ids.