Repeat pending fills when changing parameters
In the flood fill and magic wand tools, the fills will now update when
you change stuff like the size limit, tolerance, expansion etc. They
will cancel the current fill if that's still going and then repeat it
with the new parameters. This allows for quicker correction of fills
without having to cancel and redo them with other parameters.