Fix version comparison of stable against beta
It considered 2.2.2 to be older than 2.2.2-beta.1, because it would
assume that the former had a beta version of 0. That's fixed now and a
test created for that kind of stuff. Presumably this won't cause us
issues in practice, since so far the 2.2.2-beta.1 has been a server-only
release. And even if someone's bungling around on it, they'll be
notified of a beta update being present.