Collect symbols for Windows in a separate artifact
Because including them with the regular release makes it ginormous. We
now exclude the PDB files again and gather them separately. Currently
this only happens for the 64 bit Windows client, since that's really the
only thing we care to gather debug information for currently, but all
Windows targets get built as RelWithDebInfo for consistency.