Don't change fill color when switching tools
If switching from the brush to the fill tool, issuing a fill and then
applying it by switching to the eraser, the fill would change to the
eraser slot's color right before applying itself. That's fixed now by
changing the brush settings' color change signal to be a queued one and
issuing the tool change signal before the foreground change signal.