Turn too large transforms into cut and paste
The limit is at around 8K, after which the engine refuses to apply a
transform. Those now get turned into cut and paste operations and put
selection operations instead. The limit is there to avoid chugging from
ginormous transforms taking a long time, not to prevent large transforms
altogether, so this shouldn't cause issues in practice.