Allow configuring brush slot count
From 1 to 9 slots, with the eraser slot being everpresent. If only one
brush slot is configured, it gets a brush icon instead of a 1 on it.
Shortcuts for missing slots are disabled. We continue saving and loading
the eraser to and from slot 6 in the settings file for the sake of
compatibility.