Guess touch draw mode on the browser by default
So that phone users don't have to dig through the preferences to enable
touch drawing. We now guess if the user wants to draw via touch by
having received a tablet event or not. Touch drawing is the default, the
first tablet event will switch it over to pan instead. The user can
still explicitly specify what they want in the preferences.