Don't use weird recursion to refresh announcements
Because that can lead to an infinite loop that leads to the server not
feeling well. The whole reason it was implemented like this was to
bucket session announcement refreshes per list server, which is a pretty
goofy way to do it. Now it just builds a map of lists instead.