Add SignPath signing to the Actions workflow
For Windows, provided by SignPath.io and with a certificate from the
SignPath Foundation.

Only Windows client builds for stable and beta releases are signed this
way. The continuous development release, server and command-line tools
are not, since we really don't need it for those.

A link to the code signing policy is automatically prepended to the
relevant release notes in the GitHub releases pages, but at the time of
writing the link still 404s because it's not yet merged and deployed to
the website.