Make it clearer when a permission is denied
Instead of just disabling stuff or sending invalid messages that get
filtered out, we now try to give more sensible indication for when you
don't have a permission. The annotation, laser and fill tool now show a
permission denied page, the former two also show this when
annotations/layers are hidden and offer a link to re-enable them. Many
actions, like undo or paste, now let you attempt them via shortcut or
toolbars, but then show a message that you can't do that. The menu
options continue be grayed out instead. The transform toolage now also
checks if you're allowed to make a particular transform and don't make
it if you're not allowed to do so, which particularly may happen if the
user does something that upgrades a move to a put image operation.

The tool controller will also now attempt to cancel the tool multipart
operation when switching a tool and finishing the operation left it
still in a multipart state, which can happen with transforms in those
permission denied cases.

Implements #1198.