Make expanding the canvas open the resize dialog
Which will continue to let you hit the expand shortcut, but you then
have to hit return to apply it (or do whatever else in that dialog you
want to do.) This avoids the all too common accidental resizes and has
better performance than doing a bunch of separate resize steps. The
keyboard shortcuts in this dialog now also auto-repeat.

Relates to #901.