Lazy-load flood fill source only when needed
Instead of merging, flooding and gap filling the entire source area,
this is now done only when the fill needs them. This particularly speeds
up fills with large or unlimited size limits, since they no longer do a
bunch of work across that entire area in advance only to then fill a
small part of it.