Disable color wheel preview on Android and Web
It steals inputs from the main window on Android, making the color wheel
hardly usable while it's enabled. Not sure what can be done to fix this,
but Krita seems to manage in ther KoColorPopupAction.

In the browser, the window just doesn't show up at all.

On both of those platforms we'll just disable it for now.