Add color circle dock
Showing something similar to what Krita calls "artistic color selector"
and MyPaint calls "HSV/HCY wheel". Splits the three components into a
slider for the value/lightness/luminance and a circle for hue and
saturation/chroma. Can be configured to restrict the number of steps for
each of these dimensions.