Sort and expand dock toggle action titles
The names are kind of shortened to fit onto the tabs well, so now
there's a separate "full title" with a proper name, e.g. "Wheel" vs.
"Color Wheel". The toggles menu is also now sorted.

Also moves all of the dock titles into the docks themselves, before some
of them were passed via the main window and the translation was bungling
around in there with little context.