Fix effective interpolation detection
It used a malformed quad for the source, so Qt always said that there
was no way to transform it to the destination, leading to the raw
interpolation being used every time. Now it detects trivial transforms
properly and returns to nearest-neighbor interpolation as it should to
avoid pointless blur.