Build 32 bit Android against API level 23
Which is Android 6, rather than against newer versions. This is also how
we do it on F-Droid, since that gives broader compatibility. Older
Android versions stop working when we target a too high SDK, but
targeting the most recent SDK is pretty much required due to Google
blocking stuff built against older ones claiming that the application is
insecure, even if we don't use anything security-relevant in it.