Allow ignoring mouse events for freehand tool
An attempted hack around XP Pen problems when it is in relative mode. It
seems to send mouse events first and then doesn't send tablet events
when those are accepted, so this is a hack around that which will ignore
those mouse events altogether when the freehand tool is active to test
that theory. This isn't ideal like this though, since it will also
ignore mouse inputs when using canvas shortcuts and doesn't update the
outline position while the tool is active. It may get reverted or
adjusted further from here.