Allow quickly moving selections by dragging them
Reducing the amount of clicks required for such quick operations, making
it work similarly to how it did before, without requiring a double-click
to deselect. Clicking on the selection gets you full transform handles.
Dragging at the edge of a selection will move only the selection itself.
Can be toggled via a checkbox in the transform settings.