Client info and password-dependent hosting
Adds an early client info step in the login process, support of which is
indicated by the CINFO server flag. This transmits the hostname,
allowing the server to check and optionally restrict it. Enabling that
restriction locks out older clients, since they don't send this message.

This also fixes the issue of not being able to host sessions via the
browser client even if passworded sessions would be allowed. It now
allows them, erroring only if the session isn't given a password.

There's also some smaller fixes with missing state changes and returns
in error cases where that would be more proper.