Cover at least one pixel when flood filling
The fill area can end up empty due to erosion, e.g. because the user set
the selection to shrink or fill gaps. That's not sensible though, since
it just leads to nothing happening even though the user issued a fill
and prevents dragging the cursor to adjust the tolerance parameter. So
instead, we now always return at least the initial pixel filled,
appropriately masked by the selection.

Implements #1399.