Shortcut flood fills on blank layers and canvas
Users are prone to use the flood fill tool to fill the entire canvas,
which wastes a bunch of time flooding the empty area. This adds an
optimization that checks if the layer or canvas look blank and in that
case skip the flooding business altogether. The optimization will at
most check three layers when filling the canvas and only takes tile
bounds into account, since this is meant to be a quick check.

Implements #1414.