Fix up server setting when showing host dialog
When the user is coming from an earlier version of Drawpile, their
previously-selected server is now translated to the new host dialog
properly. That is, if they hosted on pub they get the pub option, if
they hosted on another server they get that one, and if they hosted on
the builtin server they get that.

This will also translate manually hosting on pub by entering its address
(or any adjacent drawpile.net address) to the actual pub option.