Explicitly check overflow when creating DP_Rect
It doesn't cause issues in practice because the resulting rectangle ends
up invalid anyway, but better to avoid the UB and explicitly return an
invalid rectangle.

On GCC and Clang, we use the builting overflow check intrinsic. On other
compilers, we do a classic bounds check. This also adds limits.h to
common.h, so that include is tossed from several other files.