Allow merging layers in more configurations
It now intuits the correct top-level id to merge into, so even if the
current id is the child of a selection, you're still allowed to merge
and it should do what you mean, rather than refusing to merge until you
move the current item to top-level. Merging a layer down into a group is
also allowed now, it leads to the group getting merged first and then
the layer gets merged into the result.