Conditionally load less optimized WASM [ci skip]
In the browser version of Drawpile. Webkit bug 284752 causes the
application to blow up when building with full optimizations, but the
relwithdebuginfo build appears to work. If we detect a user running
Safari 18 or newer, we fall back to that. Later we should cap this to
the actually affected versions, but currently it's not fixed yet, so
it's safer to always use it. We also now have to load the drawpile.js
file dynamically because of this.