Handle nested errors during socket error handling
Attempting to disconnect an erroring socket can in turn cause another
error to be emitted, which in turn will call the handler again, which
will then attempt to disconnect again until the program blows up. The
error handler now checks if it's already handling an error and bails out
in that case.

Fixes #1455.