Improve kinetic scrolling on the layer list
Instead of it fighting with dragging layers and sliding across
checkboxes when kinetic scrolling is set to left-click or touch, it now
only enables scroll on the right side of the list, minus the checkboxes.
This feels much more sane to use, dragging the layer will move it, but
dragging the blank area on the right will scroll.