Implement lightness and darkness erase blend modes
Blend modes that turn the lightness into alpha in some fashion.

Erase Lightness is a blend mode that erases according to the lightness
and opacity of the pixel. A white pixel erases, a black pixel doesn't.
Erase Darkness is the inverse of that. These can be used to use
grayscale layers for masking, which is similar to how other programs do
it, so it would possibly allow copying over a mask from there.

Lightness to Alpha and Darkness to Alpha basically erase the layer with
its own lightness/darkness. I am told this is useful for a technique
where you shade/light with black and white and then apply the colors
afterwards.