Use int to represent layer and annotation ids
Where not interacting with messages directly. This is already done in
most places in the applications, but there were some leftover places
that didn't. The idea is that we're going to upgrade layer ids to be 24
bits instead of 16 to allow users to create more than 256 layers without
needing to be operator.