Implement "clip to layer below" layer attribute
This is how some programs and PSD files represent alpha preserve. Layers
that have this attribute set will create an implicit isolated layer
group with the layer below. The layer group's properties get set to the
properties of the base layer and the base layer's properties are reset
to Normal blending at full opacity. The clipped layers are composited on
top of it with alpha preserve.

These kinds of clipped layers are saved into ORA files as the reified
layer groups described above. They get a `drawpile:clipping-group`
property attached to them, as well as the title `__clippinggroup`
followed by a number so that Drawpile can load them again. Opening them
in Krita gives the same visual result and due to the special name
Drawpile will correctly restore the clipping groups unless they're
messed with. PSD files have a clip flag on layers that's simply saved
and loaded accordingly.