Represent dptxt message bodies as JSON
Because the bespoke format it used before had round-trip problems and
failed tests because of it sometimes. Doing it this way should also make
the format more hackable and take less effort to maintain.

This also removes the need for a special name for the linear light blend
mode, which couldn't be called "krita:linear light" due to that
containing a space, which wasn't supported by the dptxt format.