Allow assigning colors to key frames
The color just gets stashed into the title as a prefix of the form
"#abcdef!", since that's a fine place for it that still has plenty of
room. We offer the same set of marker colors as Krita, with an escape
hatch of just manually typing in the color marker into the title and
getting whatever color out of it.