Include layer id in pen up command
Since it's silly to have to check every single layer whether it has a
sublayer to merge, since the brush engine knows what layer you're
drawing on. So that layer is now sent along with the message and cuts
the search short. A layer id of 0 still searches through every layer,
which is currently used for the cleanup when disconnecting from a
session.