Increase layer ids size, allow selection targets
Layer ids are now 24 bits wide in the protocol. The bottom 8 bits are
for the context id and 15 bits above that are for the layer element id.
This effectively increases the layer limit from 256 to 32767 per user.

The top bit being 1 indicates that the id is actually referring to a
selection, in which case the bottom 8 bits are still the context id, but
the 8 bits above are the selection element id. The 7 bits above are
reserved and must be zero. The special cases to make the move rect and
transform region messages refer to selections are gone now, they just
use these selection layer ids instead.

There's some UI that lets you enable drawing on selections. You can also
change the color of the selection mask now and the mask is actually
rendered as part of the canvas now, rather than being a QGraphicsItem on
top. The selection mask image is also now only requested lazily, to
avoid instantiating it over and over again when you're drawing on the
selection.

All other ids also now put the context id into the bottom 8 bits and the
element id above that for consistency, previously it was the other way
round.

The local match logic is now handled by protogen, since there's a lot
more of them now. It considers all selection ids below 128 to be
local-only, so any messages referring to them are replaced with local
match messages. We currently only use id 1, so this doesn't really
matter, but the higher ids are supposed to be used to reify selections
for masking brush strokes.

Relates to #1148, #1312 and #1396.