Make selections mask brush strokes
When drawing with a selection present, it is now synchronized
tile-by-tile to the remote. These sync selection tile don't participate
in undos and redos, so the brush engine can assume that any tile it
synchronizes will remain that way. We could potentially use a similar
technique for brush tips in the future.

Draw dabs messages will stash the reference to the selection id they are
masked by into 5 bits of their flags parameter, which is also used to
store the paint mode. This means we get 31 selections per user that we
can mask a brush stroke by, which should hopefully be enough,
considering we only use a singular selection per user currently.

To accomplish masking of indirect strokes, layer contents now carry a
reference to a layer content that acts as a transparency mask. This gets
optimized a bit by remembering which tiles are entirely opaque, which is
also used in selections. This could be used for actual transparency
masks in the future.

Relates to #1433.