Allow brushes to synchronize smudge sampling
Making them wait until the paint engine has processed their dabs before
sampling a smudge color, which fixes artifacts with some brushes. That
can be pretty slow, so it's done in a separate thread, abstracted away
into a new stroke worker structure that takes care of delegating to the
thread or pushing to the brush engine directly.