Implement OKLAB blend mode
For brushes and layers. Works similar to pigment mode in that it
converts to another color space, does the blending and then converts
back to sRGB. Since that's also slow, the brush modes are locked behind
the same permission.

This bumps the protocol minor version, due to the incompatible blend
mode.