Fix selection mask desync between different tools
Since they each use a separate instance of the brush engine, they could
clear each others's mask state and end up in a weird state where you
can't draw on some tiles because the tool thinks they're synced already.
This is fixed by tracking the last tool that fiddled with the mask and
clearing it in that case. A better solution would be to deduplicate the
brush engines though.