Allow joining sessions with minor incompatibility
Instead of outright denying clients to join sessions with a minor
protocol difference. Since their canvas is likely to desynchronize when
features are used that their client doesn't support, they are not
allowed to reset or compress the canvas. A warning is shown to them
similar to when they are in compatibility mode, the join message will
say that they're on an older client, their name in chat will be italic
and they will be shown as "Outdated" in the user list.