Allow picking socket connection type
In the host, join and browse pages of the start dialog, there's now a
button at the bottom-right that contains actions to explicitly connect
via WebSockets or TCP. The URL will automatically be adjusted
accordingly when connecting. If a new window is spawned, the connection
strategy is transmitted to it via a new command-line switch.

If TCP sockets are not compiled in (i.e. on the browser), the button
won't be there.