Implement the basics of auto-recording
Called "autosave" in the application, since that's the more familiar
term. Currently not yet enabled by default, but can be toggled for
hosted/joined sessions separately in the files settings, as well as for
individual sessions.

Auto-recording records messages into the application data directory and
takes snapshots and creates thumbnails along the way. If the application
terminates cleanly, the associated files are deleted. Otherwise, they
stick around and can be recovered by saving them elsewhere on the
system.

This is still missing a bunch of stuff, such as saving to dppr files,
setting limits so that autosaves don't get too huge from a session
running in the background, proper reporting of playback when restoring
an autosave file and more, but this is the minimum baseline for testing.

This replaces the previous autosave mechanism of periodically saving to
the last path entirely.