Add even canvas size hack (behind compile switch)
On one of my devices, the hardware renderer causes weird jaggy pixels
when the canvas widget has odd dimensions. Forcing the dimensions to be
even fixes this. That's a really dumb hack and too scary to integrate so
shortly before the 2.3 release though, so it's behind a compile-time
switch for now.